More robust video thread. Changed tutorial to image display and additional debigging of H264 frame on error

+++ src/de/yadrone/base/IARDroneLand.java
@@ -6,5 +6,5 @@ package de.yadrone.base;
  */
 public interface IARDroneLand extends IDrone {
 	
-	public void move2D(int speedX, double angular);
+	public void move2D(int speedX, float angular);
 }
